docs: migrate realtime server implementation
This commit moves the realtime server for testing WebSockets and Server Sent
Events (SSE) in htmx that was originally in the `htmx` repository to the
current repository.

Added a new `package.json` file to install `ws` dependency, since
originally it relies on the project root `package.json`. Also updated
the loading of `htmx.js` file to use CDN instead, and corrected the
relative paths for the `ws.js` and `sse.js` files.